refactor util methods
